+++ blade-egui/src/lib.rs
@@ -109,8 +109,15 @@ pub struct GuiPainter {
 impl GuiPainter {
     /// Destroy the contents of the painter.
     pub fn destroy(&mut self, context: &blade_graphics::Context) {
+        context.destroy_render_pipeline(&mut self.pipeline);
         self.belt.destroy(context);
         for (_, gui_texture) in self.textures.drain() {
+            gui_texture.delete(context);
+        }
+        for gui_texture in self.textures_dropped.drain(..) {
+            gui_texture.delete(context);
+        }
+        for (gui_texture, _) in self.textures_to_delete.drain(..) {
             gui_texture.delete(context);
         }
         context.destroy_sampler(self.sampler);